+++ src/crud.js
@@ -1,5 +1,7 @@
 /* eslint-disable prefer-destructuring */
 /* eslint-disable no-use-before-define */
+import { attachCheckboxEventListeners, attachRemoveAllEventListener } from './taskUtils.js';
+
 const taskInput = document.querySelector('.list-item');
 const tasks = JSON.parse(localStorage.getItem('tasks')) || [];
 
@@ -45,36 +47,9 @@ function renderTasks() {
     });
   });
 
-  // Checkbox input
-  const checkBoxes = document.querySelectorAll('.checkbox');
-  checkBoxes.forEach((box, index) => {
-    box.addEventListener('change', () => {
-      if (box.checked) {
-        tasks[index].completed = true;
-      } else {
-        tasks[index].completed = false;
-      }
-      localStorage.setItem('tasks', JSON.stringify(tasks));
-    });
-  });
-
-  // Remove all tasks
-  const removeAllButton = document.querySelector('.clear-all');
-  removeAllButton.addEventListener('click', () => {
-    const incompleteTasks = tasks.filter((task) => !task.completed);
-    tasks.length = 0;
-    Array.prototype.push.apply(tasks, incompleteTasks);
-
-    // Update index list
-    for (let i = 0; i < tasks.length; i += 1) {
-      tasks[i].index = i + 1;
-    }
-
-    localStorage.setItem('tasks', JSON.stringify(tasks));
-    renderTasks(); // Update the DOM
-
-    return tasks;
-  });
+  // Call the checkbox and remove all event listener functions from taskUtils.js
+  attachCheckboxEventListeners();
+  attachRemoveAllEventListener();
 }
 
 // Remove task function
